Properly reset the delay time

+++ modules/edu.gemini.seqexec.web/edu.gemini.seqexec.web.client/src/main/scala/edu/gemini/seqexec/web/client/model/model.scala
@@ -95,7 +95,7 @@ case class SequencesOnDisplay(instrumentSequences: Zipper[SequenceTab]) {
   }
 }
 
-case class WebSocketConnection(ws: Pot[WebSocket], nextAttempt: Long)
+case class WebSocketConnection(ws: Pot[WebSocket], nextAttempt: Int)
 
 object WebSocketConnection {
   val empty = WebSocketConnection(Empty, 0)
